modifies file structure for testability
