+++ src/status/helpers.js
@@ -1,0 +1,22 @@
+  const moment = require('moment');
+  
+  function occuredWithinCurrentWeek(date) {
+      /*TODO: add timezone*/
+    const now = moment();
+    const then = moment(date, "YYYYMMDD");
+    return now.format("W") === then.format("W") &&
+    (now.format("YY") === then.format("YY"));
+ }
+ 
+  function occuredWithinCurrentMonth(date) {
+    /*TODO: add timezone*/
+    const now = moment();
+    const then = moment(date, "YYYYMMDD");
+    return (now.format("M") === then.format("M")) &&
+    (now.format("YY") === then.format("YY"));
+ }
+ 
+module.exports = {
+    occuredWithinCurrentWeek,
+    occuredWithinCurrentMonth
+}
