+++ src/status/status.js
@@ -1,0 +1,6 @@
+const {
+    occuredWithinCurrentWeek,
+    occuredWithinCurrentMonth
+} = require('./helpers')
+
+module.exports = (logProvider, logger) => require('./status.factory')(logProvider, logger, occuredWithinCurrentWeek, occuredWithinCurrentMonth)
